+++ DragonShield/helpers/UserDefaultsKeys.swift
@@ -13,4 +13,6 @@ struct UserDefaultsKeys {
     static let backupLog = "backupLog"
     static let lastBackupTimestamp = "lastBackupTimestamp"
     static let databaseMode = "databaseMode"
+    static let backupDirectoryURL = "backupDirectoryURL"
+    static let backupDirectoryBookmark = "backupDirectoryBookmark"
 }